Version 0.8.5
Enable default graphics resources mechanism for fail-safe and default behavior.
Add bindless textures mananger.
Use bindless texture to create an automatic reflexion with the material of the scene background.
Crash on termination fixed.
Update documentation.
SimpleMeshResource completed.
RenderToCubemap comepleted.
Shadow mapping completed.
Re-enable/upgrade PCF filtering for shadow mapping.
CSM for directionalLight (WIP).
Update setting keys name.
Set LibSNDFile as optional.

Minor

+++ src/Graphics/Geometry/IndexedVertexResource.hpp
@@ -174,7 +174,7 @@ namespace EmEn::Graphics::Geometry
 				/* If a sub-geometry mechanism is not used, we return 0 as offset. */
 				if ( m_subGeometries.empty() )
 				{
-					return {0, static_cast< uint32_t >(m_indexBufferObject->indexCount())};
+					return {0, m_indexBufferObject->indexCount()};
 				}
 
 				if ( subGeometryIndex >= m_subGeometries.size() )
